add cite.bib files to the newly created publications
for future reference, if new publications are added through CMS portal
then a file named "cite.bib" should be created in the corresponding
folder (in parallel with the index.md file). Search for the paper
in google scholar, under the search result link there is a hyperlink
for double quotes, click that hyperlink, there is a tab named "Bibtex"
in the bottom of the widget that pops out. Select that tab, then you
will be redirected to a webpage, and you can copy the content on that
webpage to paste it into the newly created cite.bib file.